# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LEIZ STORE API
  description: |
    REST API backend leiz-store (Next.js 14 App Router + Supabase PostgreSQL).
    Spesifikasi ini diturunkan dari `docs/api-specification.md` dan kode sumber
    di `src/app/api/**`.

    ## Konvensi
    - Semua endpoint (publik, admin, system) memakai envelope standar
      `{ success, data, meta }` untuk sukses dan `{ success, error: { code, message, details } }`
      untuk gagal. List endpoint menempatkan array di `data` dan paginasi di `meta`.
    - Semua error dikembalikan via `error.code` konsisten
      (`VALIDATION_ERROR`, `UNAUTHORIZED`, `FORBIDDEN`, `NOT_FOUND`, `CONFLICT`,
      `RATE_LIMITED`, `INTERNAL_ERROR`).

    ## Autentikasi
    - **Admin:** Supabase Auth (cookie `sb-*` atau `Authorization: Bearer <access_token>`)
      ATAU legacy JWT (cookie `admin_token` atau Bearer HS256, role `ADMIN`, 24 jam).
    - **System:** header `x-admin-secret` (`DISCORD_ADMIN_SECRET`) untuk
      `POST /api/admin/payment-action`; header `x-cron-secret` / `Authorization: Bearer`
      (`CRON_SECRET`) untuk cron.
    - **Discord:** verifikasi signature ed25519 (`x-signature-ed25519`,
      `x-signature-timestamp`) dengan `DISCORD_PUBLIC_KEY`.
  version: 1.0.0
  contact:
    name: LEIZ STORE Backend
  license:
    name: Proprietary
servers:
  - url: https://leiz-store.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
...
security: []
paths:
...
  /api/orders:
    post:
      tags: [Public]
      summary: Buat order (manual payment)
      description: |
        Membuat order dengan status `PENDING_PAYMENT`. Stok dipesan secara
        atomik via RPC `create_order_atomic`. **Idempoten** — sertakan header
        `Idempotency-Key` agar retry tidak membuat duplikat (replay mengembalikan
        order yang sama). Token konfirmasi dikirim via cookie httpOnly
        `pc_<orderId>` (path-scoped), bukan di body respons.
      operationId: createOrder
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
            maxLength: 128
            example: "5e2b9f6a-3c8d-4a1e-9f2b-7c6d5e4f3a2b"
          description: |
            Kunci idempotensi checkout. Wajib UUID v4 (diverifikasi
            `validateIdempotencyKey`); nilai lain ditolak dengan 400.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
            example:
              customerName: "Budi"
              customerDiscord: "123456789012345678"
              customerIGN: "BudiGG"
              customerNotes: ""
              items:
                - { productId: "prod_abc", quantity: 2 }
              paymentMethod: "gopay"
              currency: "IDR"
      responses:
        "201":
          description: Order berhasil dibuat.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: "Cookie `payment_confirmation_<orderId>` berisi token konfirmasi (httpOnly, path-scoped)."
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/Order"
                          - type: object
                            properties:
                              manualPayment: { type: boolean, example: true }
                              replayed: { type: boolean, example: false }
              example:
                success: true
                data:
                  id: "8f3a2b1c4d5e6f7a8b9c0d1e2f3a4b5c"
                  order_number: "LZ-20260805-ABC123"
                  status: "PENDING_PAYMENT"
                  customer_name: "Budi"
                  total: 222000
                  tax: 22000
                  currency: "IDR"
                  payment_method: "gopay"
                  expiry_at: "2026-08-06T10:00:00.000Z"
                  manualPayment: true
                  replayed: false
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
...
  /api/orders/{orderId}/confirm:
    post:
      tags: [Public]
      summary: Konfirmasi transfer pembayaran
      description: |
        Pembeli mengonfirmasi transfer manual (opsional upload bukti).
        Token konfirmasi diambil dari cookie `payment_confirmation_<orderId>`
        atau body `confirmationToken`. Status berubah ke `WAITING_CONFIRMATION`.
        Rate limit: 5 percobaan/10 menit/IP per order.
      operationId: confirmTransfer
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/ConfirmTransferRequest" }
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ConfirmTransferRequest"
                - type: object
                  properties:
                    paymentProofBase64:
                      type: string
                      deprecated: true
                      description: Tidak didukung — gunakan multipart.
            example:
              buyerName: "Budi"
              buyerDiscordId: "123456789012345678"
              note: "Sudah transfer ya"
      responses:
        "200":
          description: Transfer terkonfirmasi.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message: { type: string, example: "Transfer confirmed" }
                          orderId: { type: string }
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "429":
          $ref: "#/components/responses/RateLimitedResponse"

  # ─────────────────────────────── ADMIN ───────────────────────────────
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List user (admin)
      operationId: adminListUsers
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Cari `name` / `email` (ilike).
        - name: role
          in: query
          schema: { $ref: "#/components/schemas/Role" }
        - $ref: "#/components/parameters/PageParam"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: Daftar user + info auth (last_sign_in_at, banned_until) — envelope + meta.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
                      meta: { $ref: "#/components/schemas/Meta" }
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
    post:
      tags: [Admin]
      summary: Buat user/admin
      description: |
        Membuat user di Supabase Auth + tabel `user` (password di-hash PBKDF2).
        Rollback otomatis bila salah satu langkah gagal.
      operationId: adminCreateUser
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
            example:
              email: "staff@leiz.store"
              password: "rahasia123"
              name: "Staff"
              role: "ADMIN"
              discord: "staff#0001"
      responses:
        "201":
          description: User dibuat.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            type: object
                            properties:
                              id: { type: string }
                              email: { type: string }
                              name: { type: string }
                              role: { $ref: "#/components/schemas/Role" }
                              is_active: { type: boolean }
                          message: { type: string, example: "Admin berhasil dibuat" }
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "409":
          description: Email sudah terdaftar.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: |
        Cookie session admin. Menerima `admin_token` (legacy JWT) dan cookie
        Supabase Auth (`sb-*`).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token admin — access token Supabase Auth ATAU legacy JWT HS256
        (role `ADMIN`).
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
      description: Secret `DISCORD_ADMIN_SECRET` untuk aksi verifikasi pembayaran.
    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: "Secret `CRON_SECRET` untuk endpoint cron. Alternatifnya `Authorization: Bearer <secret>`."
    discordSignature:
      type: apiKey
      in: header
      name: x-signature-ed25519
      description: |
        Signature ed25519 Discord (dengan header `x-signature-timestamp`),
        diverifikasi terhadap `DISCORD_PUBLIC_KEY`.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Nomor halaman.
...
  responses:
    ValidationErrorResponse:
      description: Input tidak valid (400).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "customerName: Name is required; items: At least one item is required"
              details: { customerName: ["Name is required"] }
    UnauthorizedResponse:
      description: |
        Tidak terautentikasi / kredensial salah (401). Catatan: sebagian
        endpoint (confirm-transfer, payment-action, payments/expire, discord)
        membungkus `ValidationError` dengan status 401, sehingga `error.code`
        yang dikembalikan bisa `VALIDATION_ERROR` meski HTTP-nya 401.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            success: false
            error: { code: UNAUTHORIZED, message: "Authentication required" }
...
    ConflictResponse:
      description: Konflik (409).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            success: false
            error: { code: CONFLICT, message: "This checkout retry key is no longer valid" }
    RateLimitedResponse:
      description: Terlalu banyak permintaan (429).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            success: false
            error: { code: RATE_LIMITED, message: "Too many requests" }
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL_ERROR]
            message: { type: string }
            details: {}
    Meta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    # ── Enums ──
    Currency:
      type: string
      enum: [IDR, USD]
      default: IDR
    Role:
      type: string
      enum: [ADMIN, CUSTOMER]
    OrderStatus:
      type: string
      enum:
        - PENDING
        - PENDING_PAYMENT
        - WAITING_PAYMENT
        - WAITING_CONFIRMATION
        - PAID
        - PROCESSING
        - COMPLETED
        - REJECTED
        - NEEDS_REVIEW
        - CANCELLED
        - FORCE_CANCELLED
        - REFUNDED
        - EXPIRED
...
    Order:
      type: object
      properties:
        id: { type: string }
        order_number: { type: string, pattern: "^LZ-[0-9]{8}-[A-Z0-9]{6}$" }
        status: { $ref: "#/components/schemas/OrderStatus" }
        customer_name: { type: string }
        customer_discord: { type: string, nullable: true }
        buyer_discord_id: { type: string, nullable: true }
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        total_usd: { type: number, nullable: true }
        currency: { $ref: "#/components/schemas/Currency" }
        payment_method: { type: string, nullable: true }
        expiry_at: { type: string, format: date-time, nullable: true }
        confirmed_at: { type: string, format: date-time, nullable: true }
        paid_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        avatar: { type: string, nullable: true }
        discord: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        is_active: { type: boolean }
        last_login_at: { type: string, format: date-time, nullable: true }
        last_sign_in_at: { type: string, format: date-time, nullable: true }
        banned_until: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    CreateOrderRequest:
      type: object
      required: [customerName, customerDiscord, items, paymentMethod]
      properties:
        customerName:
          type: string
          minLength: 1
          maxLength: 100
        customerDiscord:
          type: string
          minLength: 1
          maxLength: 100
          description: "Snowflake 17–19 digit, user#0000, atau username Discord."
        customerIGN:
          type: string
          maxLength: 100
        customerNotes:
          type: string
          maxLength: 500
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string, minLength: 1 }
              quantity: { type: integer, minimum: 1, maximum: 999 }
        paymentMethod:
          type: string
          minLength: 1
          description: "Contoh: bank_transfer, gopay, dana, qris."
        currency: { $ref: "#/components/schemas/Currency" }
    ConfirmTransferRequest:
      type: object
      required: [buyerName, buyerDiscordId]
      properties:
        buyerName:
          type: string
          minLength: 1
          maxLength: 100
        buyerDiscordId:
          type: string
          minLength: 1
          maxLength: 100
        confirmationToken:
          type: string
          description: Wajib bila token tidak tersedia di cookie `payment_confirmation_<orderId>`.
        note:
          type: string
          maxLength: 500
        paymentProof:
          type: string
          format: binary
          description: "File bukti transfer (multipart/form-data)."
...
    CreateUserRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 6 }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        discord: { type: string, nullable: true }
        phone: { type: string, nullable: true }
...